+++ packages/web-inspector/src/index.ts
@@ -5641,7 +5641,7 @@ ${prettyEvent}</pre
       .replace(/&/g, "&amp;")
       .replace(/</g, "&lt;")
       .replace(/>/g, "&gt;")
-      .replace(/\"/g, "&quot;")
+      .replace(/"/g, "&quot;")
       .replace(/'/g, "&#39;");
   }